move error overlay, clean up some style
